+++ src/test/java/com/runners/RunSmokeTest.java
@@ -19,7 +19,7 @@ import cucumber.api.junit.Cucumber;
 		features = "src/test/java/com/features", 
 		glue = "com.stepDefinitions", 
 		tags = "@DG1-13", 
-		dryRun = false, 
+		dryRun = true, 
 		monochrome = true, 
 		strict = true
 )
