+++ test-output/LoggerReport/dashboard.html
@@ -83,7 +83,7 @@
 							<div class="navbar-text nav-title flex" id="pageTitle">Dashboard</div>
 							<ul class="nav nav-xs flex-row no-border pull-right">
 <li><a href="#" class="b-l"><span class="badge badge-pull">ExtentReports</span></a></li>
-<li><a href="#" class="b-l"><span class="badge badge-pull">Aug 13, 2022 01:33:44 PM</span></a></li>
+<li><a href="#" class="b-l"><span class="badge badge-pull">Aug 13, 2022 04:34:41 PM</span></a></li>
 <li><a href="#" class="b-l"><span class="badge badge-pull">4.0</span></a></li>
 							</ul>
 						</div>
@@ -100,8 +100,8 @@
 						1
 					</div>
 					<div class="d-flex">
-						<span class="flex text-muted">Out of which 1 passed</span>
-						<span><i class="fa fa-circle-o text-success"></i> 100.00%</span>
+						<span class="flex text-muted">Out of which 0 passed</span>
+						<span><i class="fa fa-circle-o text-success"></i> .00%</span>
 					</div>
 				</div>
 			</div>
@@ -114,8 +114,8 @@
 						1
 					</div>
 					<div class="d-flex">
-						<span class="flex text-muted">Out of which 1 passed</span>
-						<span><i class="fa fa-circle-o text-primary"></i> 100.00%</span>
+						<span class="flex text-muted">Out of which 0 passed</span>
+						<span><i class="fa fa-circle-o text-primary"></i> .00%</span>
 					</div>
 				</div>
 			</div>
@@ -125,7 +125,7 @@
 						<span class="">Start</span>
 					</div>
 					<div class="py-3 text-center text-md">
-						Aug 13, 2022 01:33:44 PM
+						Aug 13, 2022 04:34:41 PM
 					</div>
 					<div class="d-flex">
 						<span class="flex"></span>
@@ -139,7 +139,7 @@
 						<span class="">Time Taken</span>
 					</div>
 					<div class="py-3 text-center text-md">
-						0h 0m 9s+29ms
+						0h 0m 0s+291ms
 					</div>
 					<div class="d-flex">
 						<span class="flex"></span>
@@ -163,14 +163,14 @@
 							<div class="col-lg-5">
 								<div class="list">
 									<div class="list-item">
-										<div><i class="badge badge-xs badge-o b-info"></i> &nbsp;Pass 100.00%</div>
+										<div><i class="badge badge-xs badge-o b-warning"></i> &nbsp;Skip 100.00%</div>
 									</div>
 								</div>
 							</div>
 						</div>
 					</div>
 					<div class="box-footer">
-						<small class="text-muted">1 passed, 0 failed and 0 skipped</small>
+						<small class="text-muted">0 passed, 0 failed and 1 skipped</small>
 					</div>
 				</div>
 			</div>
@@ -188,14 +188,14 @@
 							<div class="col-lg-5">
 								<div class="list">
 									<div class="list-item">
-										<div><i class="badge badge-xs badge-o b-info"></i> &nbsp;Pass 100.00%</div>
+										<div><i class="badge badge-xs badge-o b-warning"></i> &nbsp;Skip 100.00%</div>
 									</div>
 								</div>
 							</div>
 						</div>
 					</div>
 					<div class="box-footer">
-						<small class="text-muted">1 passed, 0 failed and 0 skipped</small>
+						<small class="text-muted">0 passed, 0 failed and 1 skipped</small>
 					</div>
 				</div>
 			</div>
@@ -213,15 +213,15 @@
 							<div class="col-lg-5">
 								<div class="list">
 									<div class="list-item">
-										<div><i class="badge badge-xs badge-o b-info"></i> 100%</div>
-										<div class="list-body"><span class="text-muted">Passed</span></div>
+										<div><i class="badge badge-xs badge-o b-warning"></i> 100%</div>
+										<div class="list-body"><span class="text-muted">Skipped</span></div>
 									</div>
 								</div>
 							</div>
 						</div>
 					</div>
 					<div class="box-footer">
-						<small class="text-muted">6 passed, 0 failed and 0 skipped</small>
+						<small class="text-muted">0 passed, 0 failed and 8 skipped</small>
 					</div>
 				</div>
 			</div>
@@ -253,13 +253,13 @@
 						</tr>
 						<tr>
 							<td>@victor</td>
-							<td>1</td>
+							<td>0</td>
 							<td>0</td>
 							<td>0</td>
 						</tr>
 						<tr>
 							<td>@DG1-13</td>
-							<td>1</td>
+							<td>0</td>
 							<td>0</td>
 							<td>0</td>
 						</tr>
@@ -286,7 +286,7 @@
 		                datasets: [{
 		                    borderColor: 'transparent',
 		                    data: [
-		                    	1, 0, 0, 0, 0, 0
+		                    	0, 0, 0, 0, 0, 1
 		                    ],
 		                    backgroundColor: [
 		                        "#00af00", "#F7464A", "#8b0000", "#ff6347", "#FDB45C", "#1e90ff"
@@ -320,7 +320,7 @@
 			                datasets: [{
 			                    borderColor: 'transparent',
 			                    data: [
-			                    	1, 0, 0, 0, 0, 0
+			                    	0, 0, 0, 0, 0, 1
 			                    ],
 			                    backgroundColor: [
 			                        "#00af00", "#F7464A", "#8b0000", "#ff6347", "#FDB45C", "#1e90ff"
@@ -355,7 +355,7 @@
 			                datasets: [{
 			                    borderColor: 'transparent',
 			                    data: [
-			                    	6, 0, 0, 0, 0, 0
+			                    	0, 0, 0, 0, 0, 8
 			                    ],
 			                    backgroundColor: [
 			                        "#00af00", "#F7464A", "#8b0000", "#ff6347", "#FDB45C", "#1e90ff"
@@ -383,7 +383,7 @@
 		})();
 	})
     var timeline = {
-"Student Admission unique admission number":6.233    };
+"Student Admission unique admission number":0.035    };
     function getRandomColor() {
         var letters = '0123456789ABCDEF';
         var color = '#';
