+++ test-output/LoggerReport/exception.html
@@ -108,12 +108,12 @@
 				<div class="p-2 px-3 text-muted text-sm">Tags</div>
 					<div class="list-item attr">
 						<div class="list-body">
-							<a href="#" class="item-title _500">org.openqa.selenium.TimeoutException</a>
+							<a href="#" class="item-title _500">java.lang.AssertionError</a>
 							<div class="item-except text-sm text-muted h-1x">
 								1 tests
 							</div>
 							<div class="tc d-none">
-								<h5 class="mb-4">org.openqa.selenium.TimeoutException</h5>
+								<h5 class="mb-4">java.lang.AssertionError</h5>
 								<table class="table">
 									<thead><tr>
 										<th id="status-col"></th>
@@ -125,10 +125,10 @@
 									</tr></thead>
 									<tbody>
 	
-<tr class="s-fail" test-id="5" parent-id="1">
+<tr class="s-fail" test-id="9" parent-id="1">
 	<td><span class="w-32 avatar circle fail"><span class="badge">fail</span></span></td>
-	<td class="_600">Then the navigation modules should be displayed <br/><span class="text-muted text-sm">Navigating to the modules</span></td>
-	<td>0h 0m 30s+248msc</td>
+	<td class="_600">Then the user will not be able to admit a student with an already existing admission number <br/><span class="text-muted text-sm">CTSMS - Student Admission unique admission number</span></td>
+	<td>0h 0m 0s+34msc</td>
 	<td></td>
 	<td>
 	</td>
@@ -137,17 +137,15 @@
 <tr class="details">
 	<td colspan="5">
 		
-		<textarea disabled class="code-block">org.openqa.selenium.TimeoutException: Expected condition failed: waiting for visibility of [[ChromeDriver: chrome on WINDOWS (edca512f004bd14feba4603465008c1c)] -> xpath: //span[normalize-space()='Student Information']] (tried for 30 second(s) with 500 milliseconds interval)
-Build info: version: '4.3.0', revision: 'a4995e2c09*'
-System info: host: 'VIC-DESKTOP', ip: '192.168.86.27', os.name: 'Windows 11', os.arch: 'amd64', os.version: '10.0', java.version: '11.0.15'
-Driver info: org.openqa.selenium.chrome.ChromeDriver
-Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 104.0.5112.81, chrome: {chromedriverVersion: 104.0.5112.79 (3cf3e8c8a07d..., userDataDir: C:\Users\Victo\AppData\Loca...}, goog:chromeOptions: {debuggerAddress: localhost:61808}, networkConnectionEnabled: false, pageLoadStrategy: normal, platformName: WINDOWS, proxy: Proxy(), se:cdp: ws://localhost:61808/devtoo..., se:cdpVersion: 104.0.5112.81, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:virtualAuthenticators: true}
-Session ID: edca512f004bd14feba4603465008c1c
-	at org.openqa.selenium.support.ui.WebDriverWait.timeoutException(WebDriverWait.java:87)
-	at org.openqa.selenium.support.ui.FluentWait.until(FluentWait.java:231)
-	at com.web.CommonUtils.waitForVisibility(CommonUtils.java:295)
-	at com.stepDefinitions.Modules.the_navigation_modules_should_be_displayed(Modules.java:25)
-	at ✽.the navigation modules should be displayed(file:src/test/java/com/features/modules.feature:8)
+		<textarea disabled class="code-block">java.lang.AssertionError: expected [The Admission No field must contain a unique value] but found [The Admission No field must contain a unique value.]
+	at org.testng.Assert.fail(Assert.java:110)
+	at org.testng.Assert.failNotEquals(Assert.java:1413)
+	at org.testng.Assert.assertEqualsImpl(Assert.java:149)
+	at org.testng.Assert.assertEquals(Assert.java:131)
+	at org.testng.Assert.assertEquals(Assert.java:655)
+	at org.testng.Assert.assertEquals(Assert.java:665)
+	at com.stepDefinitions.StudentAdmissionUniqueNumber.the_user_will_not_be_able_to_admit_a_student_with_an_already_existing_admission_number(StudentAdmissionUniqueNumber.java:60)
+	at ✽.the user will not be able to admit a student with an already existing admission number(file:src/test/java/com/features/studentAdmissionUniqueNumber.feature:12)
 </textarea>
 	</td>
 </tr>
@@ -175,7 +173,7 @@ Session ID: edca512f004bd14feba4603465008c1c
 						            </div>
 									<ul class="nav nav-xs flex-row no-border pull-right">
 <li><a href="#" class="b-l"><span class="badge badge-pull">ExtentReports</span></a></li>
-<li><a href="#" class="b-l"><span class="badge badge-pull">Aug 12, 2022 11:49:14 AM</span></a></li>
+<li><a href="#" class="b-l"><span class="badge badge-pull">Aug 13, 2022 03:51:58 PM</span></a></li>
 <li><a href="#" class="b-l"><span class="badge badge-pull">4.0</span></a></li>
 								    </ul>
 								</div>
