+++ test-output/LoggerReport/tag.html
@@ -107,9 +107,9 @@
 							<div class="tc d-none">
 								<h5>@victor</h5>
 								<div class="mb-4">
-									<span class="badge pass text-pass">1 passed</span>
+									<span class="badge pass text-pass">0 passed</span>
 									<span class="badge pass text-fail">0 failed</span>
-									<span class="badge pass text-skip">0 skipped</span>
+									<span class="badge pass text-skip">1 skipped</span>
 								</div>
 								<table class="table">
 									<thead><tr>
@@ -122,10 +122,10 @@
 									</tr></thead>
 									<tbody>
 	
-<tr class="s-pass" test-id="2" parent-id="1">
-	<td><span class="w-32 avatar circle pass"><span class="badge">pass</span></span></td>
+<tr class="s-skip" test-id="2" parent-id="1">
+	<td><span class="w-32 avatar circle skip"><span class="badge">skip</span></span></td>
 	<td class="_600">CTSMS - Student Admission unique admission number <br/><span class="text-muted text-sm">Student Admission unique admission number</span></td>
-	<td>0h 0m 6s+233msc</td>
+	<td>0h 0m 0s+35msc</td>
 	<td>		<span class="category-list attr-x">
 		<span class="category badge"><i class="fa fa-tag"></i> @victor</span>
 		<span class="category badge"><i class="fa fa-tag"></i> @DG1-13</span>
@@ -135,13 +135,17 @@
 	</td>
 	<td class="static"><i class="static fa fa-external-link"></i></td>
 </tr>
+<tr class="details">
+	<td colspan="5">
+	</td>
+</tr>
 									</tbody>
 								</table>
 							</div>
 						</div>
 						<div>
 							<span class="item-meta text-xs lt">
-							<span class="badge pass text-pass">1</span> <span class="badge fail text-fail">0</span> <span class="badge skip text-skip">0</span>
+							<span class="badge pass text-pass">0</span> <span class="badge fail text-fail">0</span> <span class="badge skip text-skip">1</span>
 							</span>
 						</div>
 					</div>
@@ -154,9 +158,9 @@
 							<div class="tc d-none">
 								<h5>@DG1-13</h5>
 								<div class="mb-4">
-									<span class="badge pass text-pass">1 passed</span>
+									<span class="badge pass text-pass">0 passed</span>
 									<span class="badge pass text-fail">0 failed</span>
-									<span class="badge pass text-skip">0 skipped</span>
+									<span class="badge pass text-skip">1 skipped</span>
 								</div>
 								<table class="table">
 									<thead><tr>
@@ -169,10 +173,10 @@
 									</tr></thead>
 									<tbody>
 	
-<tr class="s-pass" test-id="2" parent-id="1">
-	<td><span class="w-32 avatar circle pass"><span class="badge">pass</span></span></td>
+<tr class="s-skip" test-id="2" parent-id="1">
+	<td><span class="w-32 avatar circle skip"><span class="badge">skip</span></span></td>
 	<td class="_600">CTSMS - Student Admission unique admission number <br/><span class="text-muted text-sm">Student Admission unique admission number</span></td>
-	<td>0h 0m 6s+233msc</td>
+	<td>0h 0m 0s+35msc</td>
 	<td>		<span class="category-list attr-x">
 		<span class="category badge"><i class="fa fa-tag"></i> @victor</span>
 		<span class="category badge"><i class="fa fa-tag"></i> @DG1-13</span>
@@ -182,13 +186,17 @@
 	</td>
 	<td class="static"><i class="static fa fa-external-link"></i></td>
 </tr>
+<tr class="details">
+	<td colspan="5">
+	</td>
+</tr>
 									</tbody>
 								</table>
 							</div>
 						</div>
 						<div>
 							<span class="item-meta text-xs lt">
-							<span class="badge pass text-pass">1</span> <span class="badge fail text-fail">0</span> <span class="badge skip text-skip">0</span>
+							<span class="badge pass text-pass">0</span> <span class="badge fail text-fail">0</span> <span class="badge skip text-skip">1</span>
 							</span>
 						</div>
 					</div>
@@ -219,7 +227,7 @@
 									        </a>
 								        </li>
 <li><a href="#" class="b-l"><span class="badge badge-pull">ExtentReports</span></a></li>
-<li><a href="#" class="b-l"><span class="badge badge-pull">Aug 13, 2022 01:33:44 PM</span></a></li>
+<li><a href="#" class="b-l"><span class="badge badge-pull">Aug 13, 2022 04:34:41 PM</span></a></li>
 <li><a href="#" class="b-l"><span class="badge badge-pull">4.0</span></a></li>
 								    </ul>
 								</div>
